feat: change some vars locations
