+++ testutils/initialize.go
@@ -2,7 +2,6 @@ package testutils
 
 import (
 	"context"
-	"crypto/ecdsa"
 	"fmt"
 	"math/big"
 	"os"
@@ -14,15 +13,6 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/taikoxyz/taiko-client/bindings"
 	"golang.org/x/sync/errgroup"
-)
-
-var (
-	proposerPrivKey = "ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80"
-	proverPrivKey   = "59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d"
-	TestPrivKey     *ecdsa.PrivateKey
-	TestAddr        common.Address
-	ProverPrivKey   *ecdsa.PrivateKey
-	ProverAddr      common.Address
 )
 
 func init() {
@@ -64,7 +54,7 @@ func initTestAccount() (err error) {
 		panic(err)
 	}
 	TestAddr = crypto.PubkeyToAddress(TestPrivKey.PublicKey)
-	log.Info("TestAccount:", "privateKey", TestPrivKey, "address", TestAddr)
+	log.Info("Test Account:", "address", TestAddr.Hex())
 	return nil
 }
 
@@ -74,7 +64,7 @@ func initProverAccount() (err error) {
 		return err
 	}
 	ProverAddr = crypto.PubkeyToAddress(ProverPrivKey.PublicKey)
-	log.Info("Prover Account:", "privateKey", ProverPrivKey, "address", ProverAddr)
+	log.Info("Prover Account:", "address", ProverAddr.Hex())
 	return nil
 }
